Reorganize codebase: split routes, move tests, fix naming
- Move test_exporters.py and test_importers.py from root into tests/,
  removing duplicate client fixture from test_exporters.py
- Split sorting.py (1030 lines) into sorting.py (419 lines) for
  sorting/voting/labels/inclusion routes and detectors.py (515 lines)
  for detector CRUD, auto-detect, extractor CRUD, and extract routes
- Fix stale "vectorytones" references in app.py comment and
  datasets.py download filename
- Update CLAUDE.md architecture docs to reflect new route modules
  and previously undocumented test files
- Apply ruff formatting to all files
