+++ vistatotes/datasets/__init__.py
@@ -1,20 +1,26 @@
 """Dataset configuration and management."""
 
 from vistatotes.datasets.config import DEMO_DATASETS
-from vistatotes.datasets.downloader import (download_20newsgroups,
-                                            download_cifar10, download_esc50,
-                                            download_file_with_progress,
-                                            download_ucf101_subset)
+from vistatotes.datasets.downloader import (
+    download_20newsgroups,
+    download_cifar10,
+    download_esc50,
+    download_file_with_progress,
+    download_ucf101_subset,
+)
 from vistatotes.datasets.importers import get_importer, list_importers
 from vistatotes.datasets.importers.base import DatasetImporter, ImporterField
-from vistatotes.datasets.loader import (export_dataset_to_file,
-                                        load_cifar10_batch,
-                                        load_dataset_from_folder,
-                                        load_dataset_from_pickle,
-                                        load_demo_dataset, load_esc50_metadata,
-                                        load_image_metadata_from_folders,
-                                        load_paragraph_metadata_from_folders,
-                                        load_video_metadata_from_folders)
+from vistatotes.datasets.loader import (
+    export_dataset_to_file,
+    load_cifar10_batch,
+    load_dataset_from_folder,
+    load_dataset_from_pickle,
+    load_demo_dataset,
+    load_esc50_metadata,
+    load_image_metadata_from_folders,
+    load_paragraph_metadata_from_folders,
+    load_video_metadata_from_folders,
+)
 
 __all__ = [
     "DEMO_DATASETS",
