+++ vistatotes/models/embeddings.py
@@ -19,6 +19,7 @@ def embed_audio_file(audio_path: Path) -> Optional[np.ndarray]:
     Delegates to :class:`~vistatotes.media.audio.media_type.AudioMediaType`.
     """
     from vistatotes.media import get as media_get
+
     return media_get("audio").embed_media(audio_path)
 
 
@@ -28,6 +29,7 @@ def embed_video_file(video_path: Path) -> Optional[np.ndarray]:
     Delegates to :class:`~vistatotes.media.video.media_type.VideoMediaType`.
     """
     from vistatotes.media import get as media_get
+
     return media_get("video").embed_media(video_path)
 
 
@@ -37,6 +39,7 @@ def embed_image_file(image_path: Path) -> Optional[np.ndarray]:
     Delegates to :class:`~vistatotes.media.image.media_type.ImageMediaType`.
     """
     from vistatotes.media import get as media_get
+
     return media_get("image").embed_media(image_path)
 
 
@@ -46,6 +49,7 @@ def embed_paragraph_file(text_path: Path) -> Optional[np.ndarray]:
     Delegates to :class:`~vistatotes.media.text.media_type.TextMediaType`.
     """
     from vistatotes.media import get as media_get
+
     return media_get("paragraph").embed_media(text_path)
 
 
@@ -67,6 +71,7 @@ def embed_text_query(text: str, media_type: str) -> Optional[np.ndarray]:
         not registered or the model is not loaded.
     """
     from vistatotes.media import get as media_get
+
     try:
         return media_get(media_type).embed_text(text)
     except KeyError:
